[DAL] Widget cost - getFloat instead of getInt
[DAL] Widget cost - getFloat instead of getInt

+++ DAL/src/com/lge/spartan/dal/MySQLDALImpl.java
@@ -568,7 +568,7 @@ public class MySQLDALImpl implements DAL {
                 w.setDesc(res.getString(3));
                 w.setQuantity(res.getInt(4));
                 w.setStationId(res.getInt(5));
-                w.setCost(res.getInt(6));
+                w.setCost(res.getFloat(6));
                 widgetList.add(w);
             }
         } catch (Exception e) {
